Add option for skipping user xattrs. On certain file systems, such as TMPFS, user xattrs are not supported, which will cause setxattr() to fail, and current unsquashfs behavior will stop writing any xattrs from that point on.
However, it may still be desirable to write
other xattrs, such as SELinux labels, which
are in the security namespace, which
may be supoorted (again, TMPFS).

Unpacking a live image to TMPFS for a
R/W filesystem is a very common operation,
and current unsquashfs code makes it impossible
to preserve SELinux labels if user xattrs are
also present.

This patch provides a parameter to skip
the user xattrs when unpacking the squashfs.