Mksquashfs: fix rare race condition in fragment writing/duplicate check
This race condition only occurs using very small caches, produced
by setting the -mem option to a very low value, for example 20 Mbytes.

When the compressed fragment block is written to the fragment orderer
thread, the fragment_table isn't updated until it is processed by
the fragment orderer thread.  Which leaves the fragment table
not updated for the time it is in the queue and in transit to the
fragment orderer thread.

There are some front-end fragment de-duplication threads which
de-duplicate fragments (if possible) before they are processed
and de-duplicated by the main thread.  If a fragment is a possible
duplicate of a fragment contained in a previously created fragment
block, they first try to lookup the uncompressed fragment block in the
fragment block cache (which holds the last X created fragment blocks
on a freelist, before reuse).   If that lookup fails (because that
block has been reused), they then try and lookup the compressed
fragment block in the compressed block cache.  If that fails, the
data has to be read from the output filesystem.

The threads are "protected" from looking up a compressed fragment
before it is ready (i.e. still being compressed) because the
uncompressed fragment block by definition will always be found, because
it is released only once the compression has finished and the compressed
fragment block queued to the fragment orderer thread.

But the updating of the fragment table in the fragment orderer thread,
leaves a window where the uncompressed fragment block has been
released, but the fragment table hasn't been updated, which leads
to a failure in the fragment de-duplication threads using
uninitialised data.

Obviously, this will only happen if the uncompressed fragment
block has been (almost immediately) re-used within that window.
This will only occur with a very small fragment block cache of
a couple of blocks.

The fragment table should have been updated before queuing the
compressed fragment to the fragment orderer thread, and that
is what this commit does.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>