Mksquashfs: add -one-file-system-x option
This option behaves the same as -one-file-system did before
the previous bug fix.  This is added in case some people are
using and want the previous behaviour.

I could have made this option the one that behaves like
tar/rsync, but, I think doing it this way in the long term
is less likely to cause confusion and complaints.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>