Unsquashfs: add -match option
Fixes https://github.com/plougher/squashfs-tools/issues/202

-missing-symlinks will abort if any part of the extract pathname cannot
be resolved, because, obviously the pathname must be resolvable to
resolve and check symlinks.

That fact that this option is effectively "abort if any extract pathname
cannot be resolved" is not made clear in the help text/manual, because
it concentrates on symlinks.

So, add an extra option, -match which is basically -missing-symlinks,
but, which is described as above.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>