generate-manpages: Use non-builtin echo and not shell echo
Back in March (commit 338581e076b42f57b13682519b5bac8ac8a0c978) I first
hit the problem that "echo -e" is dealt with differently by
the various shell echos.  With the dash echo builtin not
recognising the "-e" and causing issues if both -e and dash
are used.

At that time I didn't want to add a dependency on bash, and instead
choose to rewrite the echos without using -e.   Unfortunately I
have used "-e" in a commit this week (because I am used to using -e
and it is easy to forget it doesn't work on dash).  So that isn't
a workable solution.

The solution chosen here is to use the non-builtin echo, which
recognises -e.  This should be available on machines running
dash.