xattrs-add: sort options for greater filesystem reproducibility
Previously the xattrs were added to each file in the order they
were specified on the command line (but in reverse order).

This means reproducing the filesystem depends on the order of
xattr-adds as well as the xattr-adds themselves.

This is considered to be an unnecessary dependency which may confuse
many people - leading to complaints there is a bug in the program.

So sort the xattr-add options so that they're always added to the
files in the same order.

Important note: only the xattr-add options are sorted, and then added
to the files after the xattrs read out of the filesystem.  This is
to not break compatibility with previous reproducible Mksquashfs
versions (4.4, 4.5 and 4.5.1)

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>