xattrs-add: add support for text encoding of value
Values can now be specified in text encoding.  This is
indicated if prefixed by either 0T or 0t.

This encoding is similar to binary encoding, except,
backslashes are specially treated, and a backslash
followed by 3 octal digits can be used to encode any
ASCII character, which obviously can be used to
encode control codes which cannot be entered on
the keyboard, e.g.

    -xattrs-add "user.comment=hello_world\012"

will set the user.comment xattr to "hello_world" with
a trailing newline.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>