Pseudo files: Add pseudo xattr definition
Add pseudo xattr definition, which adds an xattr to a particular
file.

Syntax:

-p "pathname x <name>=<val>"

where <name> is the name of the extended attribute including
prefix, and <val> is the extended attribute value.  This
can be prefixed with

0s - base64 encoded
0x - hexidecimal
0b - binary
0t - text, where backslashed octal values are treated specially

If no prefix, default is binary.

Multiple x options can be given to add multiple xattrs to a file.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>