xattrs: sort the pseudo xattrs and add in sort order
Sort the pseudo xattrs and add the global xattrs and the
pseudo xattrs to the xattr_list in sort order.

This ensures the same filesystem will be produced irrespective of
the ordering of pseudo xattr adds, and whether a particular
xattr was added by -xattrs-add or via a pseudo xattr add.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>