pseudo files: support reading edited pseudo files from stdin
Previous commits have allowed pseudo files to be output to stdout
by Unsquashfs, and to be read in from stdin by Mksquashfs.

This allows the output from Unsquashfs to be directly piped to
Mksquashfs, avoiding an intermediate file, e.g.

unsquashfs -pf - image.sqsh | mksquashfs - new.sqsh -pf - -comp zstd -b 1M

Will allow an image to be changed to zstd compression at a 1M block size
without needing it being unpacked to an intermediate filesystem or mounted.

However, this relies on Mksquashfs reading back the pseudo file in
*exactly* the same way it was written to by Unsquashfs.  This is because
stdin is non-seekable and Mksquashfs cannot read the data in a different
order to that in the pseudo file.

As long as the pseudo file or output filesystem is *unedited* that is
not a problem, as Mksquashfs is going to want the data in exactly the
way it was ordered by Unsquashfs.

If Mksquashfs, however, was told to exclude some files, then the data
will not be in the correct order, because it will have to seek to skip
over the data from the excluded files.

Additionally, if the filenames were edited in the pseudo file, this
again may make Mksquashfs want to access the data in a different order.
This is because files in Squashfs filesystems are ordered on disk in
alphabetical order, and thus are read that way from the input (in this
case the pseudo file).

If the edited filenames in the pseudo file are now in a different
order alphabetically, then the data will now be out of scan order,
for example:

The pseudo file may have the filenames: apple, blackberry, orange, and
strawberry.  The data for those files would thus be output in the order
apple, blackberry, orange and strawberry.

But if the filename orange was renamed to be banana, then the order changes
alphabetically to apple, banana (was orange), blackberry and strawberry.
The data for banana is now in the wrong scan order in the pseudo file.

To cope with this Mksquashfs must be able to scan forward when data is
out of order, buffering the intervening data.  Obviously data for a file
may now be in the buffered data, or in the pseudo file, and there
may be multiple files out of order, requiring multiple readahead buffers
to be tracked.

This commit does that.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>