unsquashfs: ensure regular file is writable before writing xattrs
When using the open system call (man 2 open) to create a regular
file, the flags argument controls how the file is opened
(read-write, write only), and the mode argument only applies
to future accesses of the newly created file.  That is we can
create a file for writing even if the mode permissions are read-only.

Unsquashfs uses that knowledge to open a file for writing and
set the permissions in the one open system call.  This is more
efficient than initially opening the file with explicit write
permissions and then resetting the permissions to read-only
afterwards, which people may do if they're unware of the above
feature of the open system call.

Unfortunately, extended attribute writing uses separate system calls,
and if the file permissions set by the open system call are read-only,
Unsquashfs will be unable as non-root user to write the external
attributes.

So if a regular file isn't user writable and it has external attributes,
explicitly set the user writable (S_IWUSR) permission to allow the
external attributes to be written, and then reset the permissions to
the correct permissions.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>