Mksquashfs: switch global_uid to be unsigned int
Rather than overloading the global_uid variable to both hold the uid
value (>= 0) and flag whether it is in operation (-1 if not), split
it into two different variables.

This allows unsigned int to be used for the value, and  avoids the messy
unsigned int conversion.

It also brings it in line with the code for -root-uid and -root-gid.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>