Don't use sigwaitinfo()/sigtimedwait() if not supported
If sigwaitinfo() and sigtimedwait() are not supported,
use sigwait() instead.

This will disable the queue/caches dump if ^\ (SIGQUIT)
is hit twice within a second.

But the queue/caches dump is still available if SIGHUP
is sent to the program.

Currently this check is applied to MAC OS X.  FreeBSD and
NetBSD appear to have these functions.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>