Unsquashfs: Add new -all-time option
Add new "-all-time TIME" option which sets the timestamp of files
to TIME, rather than the time stored in the filesystem inode.

TIME can be an unsigned 32-bit integer since the epoch of 1970-01-01
or a string value which is passed to the "date" command to parse.

Any string value which the date command recognises can be used such as
"now", "last week", Wed Feb 08 06:02:39" or "Wed Feb 15 21:02:39 GMT 2023".

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>