Squashfs: fix an edge condition in generation of pseudo files
In Squashfs-tools version 4.5, Unsquashfs gained the ability to generate
a pseudo file representation of the input Squashfs filesystem.  This
Pseudo file could be passed to Mksquashfs to generate an exact copy
of the original filesystem, perhaps with different options such
as compression algorithm, block size etc.  This was file based,
Unsquashfs output a Pseudo file and Mksquashfs read that Pseudo file.

Recently Unsquashfs has gained the ability to pipe that Pseudo file to
stdout, and Mksquashfs has gained the ability to pipe that Pseudo file
from stdin, thus allowing Squashfs filesystems to be converted without
kernel mounting or intermediate decompression.

So the following is now possible

unsquashfs -pf - image.sqsh | mksquashfs - image2.sqsh -pf -

If the same options are used, image2.sqsh should be byte-by-byte
identical to image.sqsh.

Recent testing on an Ubuntu live-cd Squashfs filesystem produced
unexpected differences in output, that is image2.sqsh was not the same as
image.sqsh.

This was tracked down to an edge condition in sparse file handling.

Mksquashfs when reading a file will detect if there are block sized
sequences of zeros, and if so it will store that block as a "sparse"
block.  That is it is flagged as containing all zeros, and so no data
needs to be stored for it.  So any file that has large sequences of zeros
is stored sparsely in the Squashfs filesystem.  This makes the filesystem
smaller.

But Mksquashfs is careful to preserve semantics of the files it reads.
It will always store a file sparsely if it can.  But, if the file was
originally (in the source filesystem) stored non-sparsely it will
appear to user-space as a non-sparse file.  Only files that were
stored sparsely when Mksquashfs read them, will be presented as sparse
files.

This avoids the situation when a deliberately stored non-sparse file
becomes a sparse file when stored in a Squashfs filesystem or when
copied out of it.

It so happens that the pseudo file representation for a "regular file
with data" lost the semantics whether the file was originally stored
non-sparsely or was originally stored sparsely.

So if a file was originally stored non-sparsely and it was then stored
sparsely in the Squashfs filesystem, the Pseudo file definition
lost that information.  With the result that the recreated filesystem
would show the file as sparse to user-space.

This is a subtle difference where files not shown as sparse in
the original filesystem, are shown as sparse in the recreated
filesystem after conversion to a Pseudo file representation.

The fix of course is to extend the Pseudo definition to preserve
the semantics that the file was stored non-sparsely in the
original source filesystem.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>