Mksquashfs: add -mem-percent option which specifies a percentage of phys RAM
With the testing of this release (and some previous releases), I have
extensively used virtual machines to do automated and manual testing
of the tools.  This allows a lot of different configurations such as
word endianess, word size (32-bit architectue), alignment, and
physical memory sizes to be tested, giving greater test coverage.

But, the use of virtual machines has shown up some limitations in the
design of how default memory usage and memory is determined and
specified in Mksquashfs.

If you're running Mksquashfs on "bare hardware" that is not under
virtualisation, the default memory use of 25% makes good sense, because
Mksquashfs is expected to co-exist with a lot of other tasks and
processes on your machine.

But if you've spun up a virtual machine to test Mksquashfs, you
probably want Mksquashfs to use all the memory you've given the
virtual machine.  That is because you've explicitly dimensioned the
virtual machine, if you didn't want that memory to be used, you wouldn't
have given it to the virtual machine ...  So the default behaviour
of Mksquashfs to be a "good citizen" and only use 25% of the virtual machine
RAM is annoying in this case.

Of course in this case you can use the -mem option to specify an
absolute value of memory to use, such as 1024M, 8G etc.  So if you've
given the virtual machine 8 Gbytes of RAM, you could specify -mem 5G,
and if you've given it 1 Gbyte of RAM, you could specify -mem 600M.

But this fixed value doesn't change if you change the amount of
RAM you've given the virtual machine.  So you can't spin up the
same virtual machine image with different amounts of RAM, say 512
Mbytes,  1 Gbyte, 8 Gbytes etc. and have the amount of memory
Mksquashfs uses adapt to the amount of RAM given.

So this commit introduces the -mem-percent option which does just that.

Specify -mem-percent 60 and Mksquashfs will always use 60% of the RAM
given to the virtual machine.  So the same image can be used with
different amounts of RAM given to the virtual machine.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>