Mksquashfs: improve -mem help text
Now that Mksquashfs has a -mem-percent option, we can remove the
printing of the current memory usage from the help text line.

Printing it here as a fixed memory size always has been rather
misleading, as many people probably won't know it is dynamic and based
on 25% of the physical RAM of the machine running Mksquashfs.  But it was
difficult to say it was 25% here because 25% (or any other percentage)
can't be specified using -mem.  So it conceptually didn't match.

Now say that explicitly on the -mem-percent help text line, because it
conceptually makes sense here.

Also take the opportunity to make the -mem help text more concise.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>