Appending: If no match found in root directory, it will be looked up again
Normally if a Pseudo file doesn't have a dev (pseudo_ent->dev == NULL),
it can be ignored at this level - that is it doesn't matter if there's
already a file of the same name.  This is because a Pseudo file
with a NULL dev doesn't really exist.  It is just a "linking"
directory to real Pseudo files in sub-directories.

But if we're appending and in the root directory, we *do*
have to check if this NULL dev Pseudo file matches a
pre-existing file in the filesystem being appended to.  This is
because we can't recurse into that directory and add any
Pseudo files to it or any sub-directories.

So this is the reason why the file lookup is done after the NULL dev
check, unless we're appending and in the root directory,

So we have two lookups, and if the first lookup returns NULL,
we'll try again in the second lookup.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>