Mksquashfs/Sqfstar: fix race condition in fragment table update
This commit fixes a regression introduced in commit 94db791f1
(August 2022).  The race condition is caused by the update
of the "fragment table" without the fragment mutex being held.

The race condition has been seen to occur with the -tailends option,
which is non-default for Mksquashfs, but, default for Sqfstar.
The effect of the race-condition is to produce a corrupted
fragment table.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>