Mksquashfs: fix race condition in duplicate check using tailend packing
If tailends are packed into fragments (-tailends option used) in rare
circumstances a race condition will occur which may lead to a
corrupted filesystem.

Specifically all the following set of circumstances need to occur for
the race condition to trigger:

1.  The file being duplicate checked is discovered to have the same
    set of blocks as another file - but with a different non-duplicate
    tailend which needs be stored.

2. The set of blocks belonging to the file is larger than could
   be fitted into the block write cache, and so a number of blocks
   have been written to the output filesystem.

3. When storing the non-duplicate tailend block, the current fragment
   is discovered to not have enough room, and so this fragment will
   be scheduled for compression and written to disk, and another
   empty fragment will be created.

When the above happens the set of blocks in this file have been discovered
to be duplicate and so don't need to be stored (circumstance 1).   But
some blocks have already been written to disk (circumstance 2).  In this
case the current write position into the output filesystem needs to be
reset (and the output file truncated if not a block device).

Now if the current fragment block is discovered to be full it will be
scheduled to be compressed and written to disk (circumstance 3).
This triggers a race condition where the fragment may be written
to disk before or after the current write position into the
filesystem has been reset (and the output file truncated if not a
block device).  But in almost all cases compression and writing of
the fragment occurs after the rest and truncation.

So the race condition is difficult to trigger and almost always
results in the correct behaviour.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>