Mksquashfs: fix rare bug in appending with -root-becomes
This commit fixes issue https://github.com/plougher/squashfs-tools/issues/262

The above issue reports that certain filesystems when appended to
using the -root-becomes option become corrupted.

This has been tracked down to an unhandled special case in the
appending code.  Specifically when the end of the root directory is
aligned to the end of a metadata block (i.e. at an 8192 byte boundary),
the computation which splits the data into a compressed part which
is placed in the directory_table, and an uncompressed part which
is placed in the directory_data_cache does not work.

This is because the code relies on last_directory_block to be pointing
to the end of the last fully filled metadata block in the directory
table.  But the code only updates this when it reads past this
into another (unfully filled) metadata block.

This commit adds code to check for the above scenario and update
last_directory_block appropriately (at this point it is
last_start_block).

Note: the scope of this commit is to fix the above issue.
Further analysis may show additional issues which will be fixed
in later commits.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>