Add ARM64 filter support.
This sets LZMA2 options pb=2, lp=2, lc=2 as the default with the
ARM64 filter as those are good for ARM64 code (all instructions
are 4 bytes and 4-byte aligned). pb=2 is a default from the preset.
lc=2 is needed because the sum lc+lp must not exceed 4 and the
default is lc=3.

The ARM64 filter requires liblzma >= 5.4.0. It will build against
older liblzma but then the filter won't be available. An alternative
could have been to use

    #ifndef LZMA_FILTER_ARM64
    #define LZMA_FILTER_ARM64 LZMA_VLI_C(0x0A)
    #endif

which would allow building against older liblzma and make the filter
available once liblzma is upgraded, without requiring a recompilation
of squashfs-tools.

The help text lists "arm64" even if built against pre-5.4.0.

Linux 6.7 will hopefully include the ARM64 filter.
https://lkml.org/lkml/2023/11/8/775

Signed-off-by: Lasse Collin <lasse.collin@tukaani.org>