Mksquashfs: fix bug in appending in less ad-hoc way
This commit fixes issue https://github.com/plougher/squashfs-tools/issues/262
but in a less ad-hoc way to the reverted previous fix.

To recap from the previous commit:

***********************************************************************
The above issue reports that certain filesystems when appended to
using the -root-becomes option become corrupted.

This has been tracked down to an unhandled special case in the
appending code.  Specifically when the end of the root directory is
aligned to the end of a metadata block (i.e. at an 8192 byte boundary),
the computation which splits the data into a compressed part which
is placed in the directory_table, and an uncompressed part which
is placed in the directory_data_cache does not work.

This is because the code relies on last_directory_block to be pointing
to the end of the last fully filled metadata block in the directory
table.  But the code only updates this when it reads past this
into another (unfully filled) metadata block.
***********************************************************************

This commit reworks the code to set last_directory_block once all the
fully filled metadata blocks have been read, and then read any
partially filled metadata block.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>