endian_compat.h: Fix endian check on FreeBSD
When adding support for XATTRs on non-Linux systems, it
was discovered the filesystem swapping code is incorrectly
compiled to swap endianness when building read_xattrs.c
on X86_64 FreeBSD.

This was tracked down to a dependency on "machine/endian.h".
This dependency has not caused issues previously on other
source files because "machine/endian.h" is indirectly included
by "sys/types.h", which is included by the other source files,
but, not by read_xattrs.c.

This dependency may or may not exist on other BSDs, and so
explicitly including it for FreeBSD only seems too restrictive, but,
always including it will also break systems without it.

The most straight-forward option is to simply include "sys/types.h",
which will fix FreeBSD and perhaps others, and will otherwise be
harmless.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>