Makefile: Add new XATTR_OS_SUPPORT definition
When XATTR support was added to Squashfs-tools I added a
Makefile definition XATTR_SUPPORT which enabled this
new code to be enabled or disabled at build time.

This option was partially introduced because many systems (i.e.
embedded systems) don't use XATTRs, and this option enabled the
extra "code bloat" to be avoided.  But it was also introduced
because many C libraries and other operating systems don't have
XATTR support, and without this option it would make Squashfs-tools
more difficult to port to non glibc based and/or non Linux systems.

But the XATTR_SUPPORT definition is an all-or-nothing option, you
either get all the XATTR code or absolutely no-XATTR code.  Back in
the beginning when XATTRs was first added this was a non-issue
because all you could do with XATTRs was read them from the source
filesystem and write them to the target filesystem, and if
the C library or operating system didn't support that, then
there was nothing for the XATTR code to do.

Since then code have been added which allows XATTRs to be added to the
Squashfs filesystem on the command line, via actions and input and
output via pseudo-files etc. which is independent of library/OS support,
and it should be possible to use these options on a system without
library/OS support.

So this commit makes the XATTR_SUPPORT definition more fine-grained
by adding a new XATTR_OS_SUPPORT option.

XATTR_SUPPORT now adds XATTR support except for low-level OS specific
read/write routines.

XATTR_OS_SUPPORT adds support to read and write XATTRs from the source
and target filesystems.

Tested on Linux and FreeBSD.

Fixes https://github.com/plougher/squashfs-tools/issues/254

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>