generate-manpages: Determine if 'sed' cmd is GNU sed, and if not try gsed
The manpage generation scripts rely on sed being GNU sed.  Check
whether 'sed' looks like GNU sed, and if not try gsed which is often
what GNU sed is named on BSD systems.

This fixes manpage generation on FreeBSD (and probably other BSDs and
MacOS, but it has only been tested on FreeBSD).

But it also fixes the fact the scripts should not rely on 'sed' being
GNU sed and they should check for this.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>