Add RISC-V filter support
This sets the LZMA2 option lp=1 by default with the RISC-V filter
for 2-byte-aligned instructions. Sometimes also pb=1 would help but
it varies so this keeps the default pb=2.

RISC-V instructions are 2-byte aligned when the C extension (the
16-bit compressed C instructions) are used. The assumption is that
usually the C extension is used. For 4-byte-aligned instructions
lp=2,lc=2 would be better.

The RISC-V filter requires liblzma >= 5.6.0. Conditional compilation
was implemented the same way as the ARM64 filter (see 33e03bdda9b86bef),
thus the help text includes "riscv" even when the filter isn't available.

Signed-off-by: Lasse Collin <lasse.collin@tukaani.org>