pseudo: don't handle pseudo xattrs here
If a pseudo_ent hasn't got a dev, then it exists either because it is
a linking node with pseudo definitions in sub-directories, or because
it is adding a pseudo xattr.

Deal with cases where it is a linking node here, but, leave cases where
it is only adding an xattr to be handled in the pseudo xattr code.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>