pseudo: add new -pseudo-dir option that automatically creates leading dirs
Originally if you created a Pseudo file with a pathname with leading
directories that were missing, it would be ignored.  This has now been
hardened so that any missing directories are now treated as fatal.

However it could be convienient for a pre-defined directory to used in
cases where a directory is missing, rather than abort the build.

This is the best of both worlds, pseudo file definitions will now be
created without having to ensure the leading directories are always
present in the filesystem.

e.g. the following failure

$ mksquashfs test test.sqsh -p "dir/dir2/file f 0777 phillip phillip echo Hello World" -quiet
FATAL ERROR: Pathname "dir" does not exist in filesystem.  Some pseudo definitions will not be created.

will be prevented with the -pseudo-dir (or -pd for short) option

$ mksquashfs test test.sqsh -pd "d 0777 phillip phillip" -p "dir/dir2/file f 0777 phillip phillip echo Hello World" -quiet
[===================================================================|] 1/1 100%
phillip@phoenix:/tmp$ unsquashfs -lls test.sqsh
drwxr-xr-x phillip/phillip          39 2024-04-07 22:59 squashfs-root
drwxrwxrwx phillip/phillip          27 2024-04-07 23:21 squashfs-root/dir
drwxrwxrwx phillip/phillip          27 2024-04-07 23:21 squashfs-root/dir/dir2
-rwxrwxrwx phillip/phillip          12 2024-04-07 23:21 squashfs-root/dir/dir2/file

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>