mksquashfs_help: tell user what commands to run to get help if invalid option found
If an invalid option has been found, then tell the user the invalid option and
what commands can be used to get help information.

This is a lot better than printing the full help text information because
it has become too long, and just swamps the output.

e.g.

% mksquashfs ff fff ff.sqsh -ff
mksquashfs: -ff is an invalid option

Run "mksquashfs -help-section <section-name>" to get help on the following sections
        SECTION NAME            SECTION
        compression             Filesystem compression options:
        build                   Filesystem build options:
        filter                  Filesystem filter options:
        xattrs                  Filesystem extended attribute (xattrs) options:
        runtime                 Mksquashfs runtime options:
        append                  Filesystem append options:
        actions                 Filesystem actions options:
        tar                     Tar file only options:
        expert                  Expert options (these may make the filesystem unmountable):
        misc                    Miscellaneous options:
        pseudo                  Pseudo file definition format:

Or run "mksquashfs -help" to get help on all the sections

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>