mksquashfs_help: update text after fixing -Wformat-security warning
Previously the help text was printed as a format string which was a
string literal, e.g.

printf("format string");

moving this text string to an array triggers the -Wformat-security
warning because an array isn't a string literal.

The fix is to add a "%s" format string where the array becomes the
format specifier argument.

A side-effect of this, however, is to make %% characters appear as
%% rather than %.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>