mksquashfs_help: add autowrap_print() function
Add function that takes a string of formatted or unformatted text, and
wraps all lines to the maximum column size, if any exceed that length.

The first line can have a series of embedded tab characters that
define a "tab stop", the position where the cursor is placed after
wrapping the first and subsequent lines.

Currently, the function is called with a maximum column size of 80, which
matches the manually wrapped help text lines.

Lateer commits will, obviously, remove the manual wrapping of the help
text lines to 80 columns, and also add code to get the width of the terminal
rather than always using 80 columns.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>