mksquashfs_help: Get window column width and use that
Get the column width and use that when printing the text output.

Remember *do not* pipe the output of Mksquashfs to less (or other pagers),
because you'll get the standard 80 columns, because the pipe means the
output is not to a terminal ...

The next commits will add code to output to a pager (like man does).

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>