mksquashfs_help: handle missed case in autowrap_print()
Further testing showed up another corner case - if a word went over
the column width by 1 character and was then followed by a new line
('\n'), the code would not wrap the line, and instead print it as
it was.

The culprit was the loop advanced cur even if we were at the column
width, which it shouldn't do.  This confuses the code afterwards
that checks for new line, because it thinks the line is OK to print as
is.

Cur being incremented even if we were at the column width is ultimately
also the reason why the routine has to subtract 1 (cur - 1) if a
word is too big to fit into the column width, and so that work around
can be removed too.  This leaves the code that increments cur if no
progress can be made (cur - sol == 0).

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>