mksquashfs_help: determine whether /usr/bin/pager is less, more or unknown
/usr/bin/pager is a generic pager set up by many Linux distributions
(alternates mechanism) to point to the best pager available, or one
manually chosen by the user.

The general idea is any applications should respect such choices, and
use /usr/bin/pager rather than /usr/bin/less (or /usr/bin/more) directly.

This is a good idea, but, it can obfusticate what pager will actually be used.

This becomes important if you want to pass options to the pager, where
the options are pager specific.  In this case I want to pass
"--quit-if-one-screen" to less, and "--exit-on-eof" to more.

So I need to determine if /usr/bin/pager is less, more or unknown.

There are two ways of doing this, one more fool proof than the other :-

1. Follow the /usr/bin/pager symbolic links and see where they point to.

2. Run --version on /usr/bin/pager and have a look at the output.

It should be obvious the second way is more fool proof.  This is because
the first way relies on symbolic links having been made, and it won't
work if the file has been hard-linked or copied.  Also many basic Linux
distributions (such as embedded systems) have their own basic pagers which
are named less or more when they're not.  Additionally with BusyBox or
something similar, all commands are symbolic links which ultimately point
to Busybox.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>