mksquashfs: Use all bits in inode number when computing hash value
Issue https://github.com/plougher/squashfs-tools/issues/281 reports a
scenario with a network filesystem where the inode numbers do not differ
in the lowest 16 bits.

This causes an issue with inode number lookup to determine if this inode
number has already been seen, and hence we have a hardlink.

The problem is that the code uses the lowest 16 bits of the inode number
as a hash table index.  But because these values do not differ in this
scenario, it results in all the inodes hashing to the same index, which
results in an extremely slow and inefficient lookup.

The fix is to use all the inode number bits in the computation of the hash.

The inode number of 64 bits (*) is split into 4 16-bit pieces, which
are added together and the lowest 16 bits are used as the hash value.

(*) Some older systems may still have a userspace with 32-bit inode
numbers, in which case the 2 redundant adds with be optimised out.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>