sqfstar: deal better with missing option arguments
Missing arguments to options cause two parsing issues:

1. The total number of arguments on the command line will be less
   than expected.
2. The following option/filename will be evaluated in place of
   the missing argument and will most likely cause a parse
   error.

The fact is these two parse failures/rules are mutually exclusive,
that is the command line fails both rules, but, it in practice it will
fail on whichever is checked first.

It so happens in the parser there is pre-scan of the command line where
it discovers the number of arguments are less than expected, before
the actual options are evaluated.

So for example it did this

$ sqfstar -b output.sqsh
sqfstar: fatal error: no arguments specified on command line

This commit moves the check to ensure the options are evaluated
first, which should generally give a more meaningful error message.

For example

$ sqfstar -b output.sqsh
sqfstar: -b invalid block size

This commit also improves the error message when the first
rule applies and the second doesn't:

sqfstar: fatal error: no arguments specified on command line

becomes

sqfstar: fatal error: no output filesystem specified on command line

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>