unsquashfs: rmdir() directory rather than unlink() with -force
This fixes issue https://github.com/plougher/squashfs-tools/issues/282

When extracting symlinks if -force is used Unsquashfs will try to
unlink() the destination file first.  However if that file exists
and it is a directory, the unlink() will fail, leading to the
subsequent failure of the symlink() call because the file already
exists.

This commit lstats() the file first, and if it exists, calls the
appropriate unlink() or rmdir() depending on the file type.

Rmdir() will succeed if the directory is empty, which does not
produce any more safety/security risks than currently if it is
not a directory.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>